fix review comments

+++ internal/business/service/business/service.go
@@ -16,7 +16,7 @@ type businessRepository interface {
 	GetPostPhotos(ctx context.Context, postID int64) ([]string, error)
 	CheckOwnership(ctx context.Context, userID string, unitID int) error
 	CreatePost(ctx context.Context, userID string, unitID int, description string) (*entity.Post, error)
-	InsertPostImages(ctx context.Context, postID int, URLs []string) error
+	InsertPostImages(ctx context.Context, postID int64, URLs []string) error
 	GetPosts(ctx context.Context, limit, offset int) ([]entity.Post, error)
 	GetPostByID(ctx context.Context, postID int64) (*entity.Post, error)
 }
